+++ src/protobuf/input/complex/index.ts
@@ -1,0 +1,2 @@
+export { default as MapInput } from './Map';
+export { default as RepeatedInput } from './Repeated';
